+++ modules/jellyfin/plugins/resolvePlugins.nix
@@ -48,11 +48,22 @@ let
         builtins.readFile (
           builtins.fetchurl {
             inherit (repo) url;
-            sha256 = builtins.convertHash {
-              inherit (repo) hash;
-              hashAlgo = "sha256";
-              toHashFormat = "nix32";
-            };
+            sha256 =
+              let
+                inherit (repo) hash;
+                isSha256SRI = hash: (builtins.match "sha256-[a-zA-Z0-9+/]{43}=" hash) != null;
+              in
+              if isSha256SRI hash then
+                if builtins.convertHash or null == null then
+                  throw "Hash is in SRI format, but builtins.convertHash is not available in this version of Nix. Please provide the hash in nix32 format instead."
+                else
+                  builtins.convertHash {
+                    inherit hash;
+                    hashAlgo = "sha256";
+                    toHashFormat = "nix32";
+                  }
+              else
+                hash;
           }
         )
       );