+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- `nixflix.lib.jellyfinPlugins.fromRepo`'s `hash` parameter now supports nix32 format as well ([#202](https://github.com/kiriwalawren/nixflix/pull/202)).
+  - It also throws if the user's system does not have `builtins.convertHash` implemented
+
 ### Fixed
 
 - Prowlarr indexer `apiKey` inconsistency ([#210](https://github.com/kiriwalawren/nixflix/210)